requirements.txt
python-dotenv==1.0.1
numpy==1.26.4
jq==1.6.0
openai==1.14.1
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langchain_experimental==0.0.54
langsmith==0.1.27
faiss-cpu==1.8.0
ragas==0.1.4
chromadb==0.4.24
wandb==0.16.4
tiktoken==0.6.0
fastapi==0.110.0
transformers==4.38.2
huggingface-hub==0.21.4
tokenizers==0.15.2
cohere==4.56